Longest Continuous Increasing Subsequence Leetcode
Recall the problem, we need to find "the maximum length of a non-empty substring that contains only one unique character".

In other words, we need to find the Longest Substring with the same characters.

We can iterate over the given string, and use a variable count to record the length of that substring.

When the next character is the same as the previous one, we increase count by one. Else, we reset count to 1.

With this method, when reaching the end of a substring with the same characters, count will be the length of that substring, since we reset the count when that substring starts, and increase count when iterate that substring.

Therefore, the maximum value of count is what we need. Another variable is needed to store the maximum while iterating.